requirements.txt
flask
torch
openai-whisper
transformers
sentencepiece
librosa
soundfile
numpy
